merge: include classes even if there is no INCLUDE statement
I create reports that just do not include any other report but use a
class infrastructure I created in the package. I am not sure it is the
good approach how to develop ABAP reports but it works for me.

I decided to change the logic of merge which includes public classes
before the first INCLUDE statement in the way that public classes
are included right after the REPORT statement.

The old approach did not included any class to my reports because in
there is no INCLUDE statement in my reports.

To be honest, I am not sure what was the reason behind including public
classes before the first INCLUDE statement and not earlier.

Therefore, I consider this commit DANGEROUS.

If this approach does not work, I propose to define a new pragma.

----
v2:

Include the global classes after initial comments and white space as
requested by @larshp in #54.

----