pragma: fix a typo in result processing
I believe we should check the variable presult and not the variable
result, because when a pragma has no result we should not try to add the
nothing into the total result.

However, I am not even sure we should copy the pragma's definition into
result in the case where the pragma has no outcome. But I didn't remove
those statements as the pragma line can help users to figure out
problems.