wip: resolve pargmas as Pre-Processor
The newly added CUA include into a local implementations
has caused me some headaches. I wanted to do the single file merge
but that does not work because we need to load the global class.

I used to be a C/C++ programmer and we used to stop compilation after
preprocessing (gcc -E) - it replaces all macros and so on.

I consider all abapmerge pragmas as macros on steroids and thus
stopping after processor makes sense to me.

I am not sure about how to save the results.